# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Daftar Properti API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tele-app/properties": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Get list of property",
                "description": "Returns list of property",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search property by keyword",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "price[min]",
                        "in": "query",
                        "description": "Minimum price",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "price[max]",
                        "in": "query",
                        "description": "Maximum price",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "propertyType",
                        "in": "query",
                        "description": "Property type",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/PropertyType"
                        }
                    },
                    {
                        "name": "bedroomCount",
                        "in": "query",
                        "description": "Bedroom count",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bedroomCount[min]",
                        "in": "query",
                        "description": "Minimum Bedroom count",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bedroomCount[max]",
                        "in": "query",
                        "description": "Maximum Bedroom count",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bathroomCount",
                        "in": "query",
                        "description": "Bathroom count",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bathroomCount[min]",
                        "in": "query",
                        "description": "Minimum Bathroom count",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bathroomCount[max]",
                        "in": "query",
                        "description": "Maximum Bathroom count",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "lotSize[min]",
                        "in": "query",
                        "description": "Minimum lot size",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "lotSize[max]",
                        "in": "query",
                        "description": "Maximum lot size",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "buildingSize[min]",
                        "in": "query",
                        "description": "Minimum building size",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "buildingSize[max]",
                        "in": "query",
                        "description": "Maximum building size",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "ownership",
                        "in": "query",
                        "description": "Ownership",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/PropertyOwnership"
                        }
                    },
                    {
                        "name": "carCount",
                        "in": "query",
                        "description": "Car count",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "carCount[min]",
                        "in": "query",
                        "description": "Minimum Car count",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "carCount[max]",
                        "in": "query",
                        "description": "Maximum Car count",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "electricPower",
                        "in": "query",
                        "description": "Electric Power",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort By",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/ListingSort"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Order By",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "properties": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Property"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tele-app/properties/{id}": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Get property by id",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Property Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Property"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Property not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Property not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "VerifyStatus": {
                "description": "Verification status",
                "type": "string",
                "enum": [
                    "on_review",
                    "approved",
                    "rejected"
                ],
                "example": "approved"
            },
            "FacingDirection": {
                "description": "Facing Direction",
                "type": "string",
                "enum": [
                    "unknown",
                    "north",
                    "east",
                    "south",
                    "west",
                    "northeast",
                    "southeast",
                    "southwest",
                    "northwest"
                ],
                "example": "east"
            },
...
            "ListingSort": {
                "description": "Sort Listing By",
                "type": "string",
                "enum": [
                    "price",
                    "bedroomCount",
                    "bathroomCount",
                    "lotSize",
                    "buildingSize"
                ],
                "example": "price"
            },
            "ListingType": {
                "description": "Listing type",
                "type": "string",
                "enum": [
                    "unknown",
                    "sale",
                    "rent"
                ],
                "example": "house"
            },
            "PropertyOwnership": {
                "description": "Property ownership/certificate",
                "type": "string",
                "enum": [
                    "unknown",
                    "shm",
                    "hgb",
                    "strata",
                    "girik"
                ],
                "example": "shm"
            },
            "PropertyType": {
                "description": "Property type",
                "type": "string",
                "enum": [
                    "unknown",
                    "house",
                    "apartment",
                    "warehouse",
                    "shophouse",
                    "land",
                    "villa"
                ],
                "example": "house"
            },
            "Listing": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "sourceText": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "propertyType": {
                        "$ref": "#/components/schemas/PropertyType"
                    },
                    "listingType": {
                        "$ref": "#/components/schemas/ListingType"
                    },
                    "listingForSale": {
                        "type": "boolean"
                    },
                    "listingForRent": {
                        "type": "boolean"
                    },
                    "address": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "integer"
                    },
                    "rentPrice": {
                        "type": "integer"
                    },
                    "lotSize": {
                        "type": "integer"
                    },
                    "buildingSize": {
                        "type": "integer"
                    },
                    "carCount": {
                        "type": "integer"
                    },
                    "bedroomCount": {
                        "type": "integer"
                    },
                    "bathroomCount": {
                        "type": "integer"
                    },
                    "floorCount": {
                        "type": "integer"
                    },
                    "electricPower": {
                        "type": "integer"
                    },
                    "viewCount": {
                        "type": "integer"
                    },
                    "matchFilterCount": {
                        "type": "integer"
                    },
                    "facing": {
                        "$ref": "#/components/schemas/FacingDirection"
                    },
                    "ownership": {
                        "$ref": "#/components/schemas/PropertyOwnership"
                    },
                    "verifyStatus": {
                        "$ref": "#/components/schemas/VerifyStatus"
                    },
                    "city": {
                        "type": "string"
                    },
                    "pictureUrls": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uri",
                            "example": "https://example.com/image.jpg"
                        }
                    },
                    "coordinate": {
                        "properties": {
                            "latitude": {
                                "type": "integer"
                            },
                            "longitude": {
                                "type": "integer"
                            }
                        },
                        "type": "object"
                    },
                    "contact": {
                        "properties": {
                            "name": {
                                "type": "string"
                            },
                            "phoneNumber": {
                                "type": "string"
                            },
                            "company": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "user": {
                        "properties": {
                            "name": {
                                "type": "string"
                            },
                            "profilePictureURL": {
                                "type": "string"
                            },
                            "city": {
                                "type": "string"
                            },
                            "company": {
                                "type": "string"
                            },
                            "description": {
                                "type": "string"
                            },
                            "phoneNumber": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "userCanEdit": {
                        "type": "boolean"
                    },
                    "isPrivate": {
                        "type": "boolean"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Property": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "sourceText": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "propertyType": {
                        "$ref": "#/components/schemas/PropertyType"
                    },
                    "listingType": {
                        "$ref": "#/components/schemas/ListingType"
                    },
                    "listingForSale": {
                        "type": "boolean"
                    },
                    "listingForRent": {
                        "type": "boolean"
                    },
                    "address": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "integer"
                    },
                    "rentPrice": {
                        "type": "integer"
                    },
                    "lotSize": {
                        "type": "integer"
                    },
                    "buildingSize": {
                        "type": "integer"
                    },
                    "carCount": {
                        "type": "integer"
                    },
                    "bedroomCount": {
                        "type": "integer"
                    },
                    "bathroomCount": {
                        "type": "integer"
                    },
                    "floorCount": {
                        "type": "integer"
                    },
                    "electricPower": {
                        "type": "integer"
                    },
                    "facing": {
                        "$ref": "#/components/schemas/FacingDirection"
                    },
                    "ownership": {
                        "$ref": "#/components/schemas/PropertyOwnership"
                    },
                    "verifyStatus": {
                        "$ref": "#/components/schemas/VerifyStatus"
                    },
                    "city": {
                        "type": "string"
                    },
                    "pictureUrls": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uri",
                            "example": "https://example.com/image.jpg"
                        }
                    },
                    "coordinate": {
                        "properties": {
                            "latitude": {
                                "type": "integer"
                            },
                            "longitude": {
                                "type": "integer"
                            }
                        },
                        "type": "object"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "listings": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Listing"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    }
}